backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.0
pyarrow==15.0.0
sas7bdat==2.2.3
pyspark==3.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1 